Add ConnectionStream and part of LocalChatClient
I'm not entiraly sure that this will work with a single protocol for
both client and user but it should. I believe the best way to implement
point-to-point communication will be via protocol which works for both
sides of cummunication. Not sure how to proceed between steps between
client and server though.
It seems like there should be individual states for client and server
but maybe there is a better solution? Maybe client doesn't even need to
follow any protocol. Client simply outputs data to server and server
then proceed this data...